requirements.txt
 PyNaCl
 discord.py
 python_dotenv